+++ client/CHANGELOG.md
@@ -37,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Sync nodes `select!`;
 - `Response::{HexToBech32(), AliasIdToBech32()}` from message interface;
+- `ClientMinerBuilder, ClientMiner`;
 
 ## 2.0.1-rc.4 - 2022-11-22